feat: Improve blockdim optimization script substantially
This commit makes the process of generating the optimal blockdims more
convenient by not having to manually change the number of GPUs included.
Crucially, it exits the process if an error is detected. Previously,
manual detection of errors was required.

TODO: This commit does not optimize deep halos yet. It only works for
HALO_DEPTH=1

+++ laplace2d.cu
@@ -34,7 +34,7 @@ int main(int argc, const char **argv) {
 
     initialize_host_region(d_ref);
 
-    int size = BYTES_PER_GPU;
+    unsigned long size = BYTES_PER_GPU;
     if (NGPUS>1) size+=HALO_BYTES;
 #pragma omp parallel for
     for (int i = 0; i < NGPUS; i++) {
